extend ffprobe subprocess timeout
the check valid file function previously had a 0.1 second timeout. This would throw an exception if ffprobe would not spawn fast enough.

This commit extends the timeout to 5 seconds in the hope that this exception will not be thrown as often.

In the future I may remove the timeout entirely as it seems that it isn't necessary for proper script operation.